Refactor archived game, add make pgn test
